requirements.txt
fastapi==0.65.1
pydantic==1.8.2
python-dotenv==0.18.0
requests==2.25.1
starlette==0.14.2
urllib3==1.26.4
uvicorn==0.13.4
sqlalchemy==1.4.22
pytz==2021.1
contextvars==2.4
redis==3.5.3
celery==5.1.2
psycopg2==2.8.6